scripts/gdb: Add meminfo command
Provide an equivalent of /proc/meminfo which should be available from
core dumps, or crashed kernels. This should allow a debugger to identify
if memory pressures were applicable in the instance of their issue

Signed-off-by: Kieran Bingham <kieran.bingham@linaro.org>

---

I would love it if we could reduce some of the calls to parse_and_eval,
-> need to find a way to index on the values of the ENUM's instead of
calls for each one.
Then we could read in the struct and index from there.