squash! scripts/gdb: Add meminfo command
Testing lx-meminfo found calling nr_free_highpages() crashed the target.

Instant dismisal! (This wasn't supposed to be in there anyway, it's a
leftover that I hadn't got to cleaning up yet - but it could be quite a
large implementation)

This actually makes me want to drop lx-meminfo for now - and let the
other patches proceed without for now - and if this development
improves, then perhaps it could make it in. But for now - it's
information that isn't accurate - and that's just worse than broken!