fdp1: Convert to OPMODE_INTERRUPT time driven process
Setting the OPMODE to INTERRUPT means that after VPERIOD expires, if the
frame has not completed, the FDP will trigger a new VSYNC and set the
ERROR bit/interrupt as well.

For us this means we can see that the IP is actually clocking, even if
it does no work - and means we can put in the error handling for corrupt
frames as well.