squash! fdp1: Support FCP using RuntimePM hooks
You'll note that this commit contains two implementations - (the 'right
one, disabled' and the 'wrong one, enabled')

I expected that the runtime_suspend hooks would allow me to make extra
calls as neccessary and the runtime_pm framework would continue to do
the clock management of this driver for me.

However, when the .pm structure is enabled with the SET_RUNTIME_PM_OPS()
the clk_enable on my FDP1 device seems to get lost and the hardware
registers are not readable.

This squash commit puts in a workaround - but we would prefer a correct
solution of course