fdp1: try_fmt: Support deprecated fourcc formats
Gstreamer recognises that we support XRGB555, and converts that to it's
internal RGB15 format. Unfortunately, when it then tries to request
buffers on that format, it converts to RGB555 (the deprecated form) and
fails to enumerate the format successfully, or retrieve buffer sizes.

Laurent's backwards compatibility from his __vsp1_video_try_format()
is a very good fix for this.