Revert "scripts/gdb: add a Radix Tree Parser"
This reverts commit e127a73d41ac ("scripts/gdb: add a Radix Tree Parser")

The python implementation of radix-tree was merged at the same time as
the radix-tree system was heavily reworked from commit e9256efcc8e3
("radix-tree: introduce radix_tree_empty") to 3bcadd6fa6c4 ("radix-tree:
free up the bottom bit of exceptional entries for reuse") and no longer
functions, but also prevents other gdb scripts from loading.

This functionality has not yet hit a release, so simply remove it for
now

Signed-off-by: Kieran Bingham <kieran@bingham.xyz>
--

Due to unfortunate timing, the radix tree python implementation was
merged at the same time as the radix-tree itself was heavily rewritten.
Unfortunately I have not been able to find enough time to rewrite the
python implementation in this rc-window, and as the radix-tree.py has
not yet made it out to a release in a working state, at this point the
best thing to do is to revert it. (It currently causes breakage for the
other gdb scripts

I will then intend to resubmit a new implementation for a later release,
and also we can plan an automated test service for the scripts to catch
these issues earlier.