smiapp: Add MUX and metadata sub-devices for embedded data
Add two new sub-devices to SMIA compliant sensors that support embedded
data.

A "metadata" sub-device is added. The sub-device is a source for metadata.

A "MUX" sub-device is added in order to operate as a multiplexer for image
and metadata. VIDIOC_SUBDEV_G_ROUTING IOCTL is used to obtain the routing
information inside the MUX.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>