uvcvideo: Support asynchronous URB completions
The URBs are submitted back to the HCD device upon completion of the
uvc_video_complete() call.

In preparation for the decode operations performing asynchronous
actions, a reference count is added to the uvc_urb structure.

This reference count allows us to adapt the URB handling to return the
URB at a zero refcount. The count is not used to 'free' the buffer.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>