arm64: dts: renesas: salvator-x-max9286: Specify both the MAX9271 and OV10635 I2C addresses
The rdacm20 driver will set the sensor I2C address to a unique address
instead of relying on I2C address translation. For this to work, it
needs to know the desired unique I2C address for the sensor. As there's
no facility in the kernel to allocate a unique I2C address, and as
implementing such a facility would be difficult (if possible at all),
provide it in DT.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>