v4l: max9286: Configure link combination order
The MAX9286 contains a frame combination engine that interleaves lines
from all the enabled links and supports a configurable link order.
While the datasheet doesn't clearly describe how it operates, the
combination engine seems to require all enabled links to have a
configured order starting at 0 without any gap.

The default configuration assigns orders 0 to 3 to links 0 to 3. This
works fine when the enabled links start are 0 and are contiguous, but
results in not image being received at the MAX9286 output when enabling,
for instance, link 1 only or links 1 and 2.

To solve this, configure the link combination order with a valid value
based on the enabled links. The value is fixed as we don't need to
select a particular link order for now, and is thus precomputed.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>