media: i2c: max9286: Characterize FSYNCLOCK delay
It should take up to 6 complete frame periods to reach FSYNC locking
with all enabled streams. Modify the existing delay and sleep time in
the loop monitoring fsync lock to have it sleep for a duration comparable
to one third of a frame period at 30fps, and loop for a total of 12
complete frames time before returning error.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>