media: i2c: max9286: close mux channels
We can't implement the i2c_mux deselect directly as that will overload
our I2C bus with extraneous operations to disable and re-enable the
channel on every register write.

Instead, provide a helper to 'deselect' or 'close' any active mux
channel, and utilise the helper after key operations.

To match the disabling of the reverse channel, also disable the forwards
channel, ensuring that any channel's communication is only available
when that mux channel is selected.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>