arm64: dts: renesas: salvator-x-max9286: Allow flexible configuration of connected cameras
Use the preprocessor to simplify the process of defining which cameras
are connected to the Maxim-Expansion board.

The definition of the camera should define it's compatible string to be
passed to the linux driver framework.

This currently does not affect register or address settings and could be
extended to do so in the future.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>