v4l: max9286: Skip non-connected remote endpoints
We cannot assume that if we have a local endpoint, it is connected to a
remote one. Currently the return value of
fwnode_graph_get_remote_endpoint() goes unchecked.

If there is no remote endpoint, skip this local one from our notifiers.
This could be considered an error, but it is valid DT to specify the
endpoint without connecting it to a remote one. Therefore skipping this
local node seems the most appropriate action to take.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>