device property: Add .graph_get_endpoint_by_id()
The fwnode_graph_ library lacks a function to get an endpoint located at
a precise (port, endpoint) position. As v4l2-async moves towards full
endpoint matching, complex devices with multiple ports/endpoints may
want to register a specific endpoint as the one to be matched on by the
v4l2 framework.

Add to fwnode_operations structure a new member that realizes this.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>