v4l: i2c: max9286: Register subdevice on endpoint
v4l2-async framework assigns to subdevice's fwnode the fwnode associated
with device node when the subdevice is registered with fwnode field not set.
This breaks the matching procedure between rcar-csi2 registered notifier and
max9286 subdevice, as the first one expects to match the subdevice endpoint.

Fix this registering the fwnode associated with the endpoint in max9286
subdevice.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>