v4l: i2c: rdacm20: Register subdevice on endpoint
v4l2-async framework assigns to subdevice's fwnode the fwnode associated
with device node when the subdevice is registered with fwnode field not
set.
As max9286 registered notifier expects to match the fwnode associated
with the remote endpoint, fix this by assigning the subdevice fwnode
with the endpoint of rdacm20 device node.

Signed-off-by: Jacopo Mondi <jacopo+renesas@jmondi.org>