max9286: start and stop all cameras in s_stream()
To be able to capture simultaneously from multiple all cameras must be
started at the same time. Loop over all cameras in s_stream() and
start/stop them. In the start case the check for FSYNC will make sure
all cameras are in sync before s_stream() returns and indicate to the
caller that everting is ready.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>