max9286: add 'streaming mode' for the i2c mux
When staring multiple cameras the forward and reverse channels to all
cameras must be enabled. This patch adds a hackish 'streaming mode' for
the i2c mux select functions that enables these channels whenever we are
starting a stream and restores the old behavior when not streaming.

This patch should most likely be reworked into something prettier, but
for now it dose the job and demonstrates that we can capture multiple
simultaneously streams.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>