max9286: take route configuration into account when configuring links
This fixes a bug in max9286_get_routing() where the
loop assumed all connected cameras where all in sequence and gaps in the
configuration would lead to erroneous behavior.

It also adds a s_routing() operation which should be able to control how
many of the connected cameras are intended to be part of a capture.

Please note this is not meant to go into the gmsl/base branch. It is
only compile tested and the bug fix should be squashed into the original
patch which introduced the error. This patch is only posted as a preview
as an alternative to the max9286 single_source control.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
[Kieran: Convert to for_each_source; fix typo]
Tested-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>